moved tests from test_basic.py to tests/test_crash_similarity.py, added setup on testCrashSimilarity, left only test.json on paths and unzip json(easy to edit)
