+++ tests/test.json
@@ -1,0 +1,6 @@
+{"proto_signature": "mozilla::gfx::AutoRestoreTransform::~AutoRestoreTransform | gfxFont::DrawGlyphs | gfxFontMissingGlyphs::DrawMissingGlyph", "uuid": "25dd18ec-f511-4cbd-ae62-118a82161107", "signature": "mozilla::testZeroCoherence"}
+{"proto_signature": "mozilla::gfx::AutoRestoreTransform::~AutoRestoreTransform | gfxFontMissingGlyphs::DrawMissingGlyph | gfxFont::DrawGlyphs", "uuid": "25dd18ec-f511-4cbd-ae62-118a82161108", "signature": "mozilla::testZeroCoherence"}
+{"proto_signature": "mozilla::gfx::AutoRestoreTransform::~AutoRestoreTransform | gfxFontMissingGlyphs::DrawMissingGlyph", "uuid": "25dd18ec-f511-4cbd-ae62-118a82161109", "signature": "mozilla::testZeroCoherence"}
+{"proto_signature": "mozilla::gfx::AutoRestoreTransform::~AutoRestoreTransform | gfxFontMissingGlyphs::DrawMissingGlyph | gfxFontMissingGlyphs::DrawMissingGlyph", "uuid": "25dd18ec-f511-4cbd-ae62-118a82161109", "signature": "mozilla::testOrdem1"}
+{"proto_signature": "mozilla::gfx::AutoRestoreTransform::~AutoRestoreTransform | gfxFont::DrawGlyphs | gfxFontMissingGlyphs::DrawMissingGlyph", "uuid": "25dd18ec-f511-4cbd-ae62-118a82161109", "signature": "mozilla::testOrdem2"}
+{"proto_signature": "mozilla::gfx::AutoRestoreTransform::~AutoRestoreTransform | gfxFontMissingGlyphs::DrawMissingGlyph | gfxFont::DrawGlyphs", "uuid": "25dd18ec-f511-4cbd-ae62-118a82161109", "signature": "mozilla::testOrdem3"}
